Add a service to handling interfacing with vertex auth api
