+++ src/app/components/map/nav-bar/nav-bar.component.html
@@ -22,7 +22,7 @@
     class="spacing"
     color="basic"
     mat-fab>
-    <mat-icon>{{ isLoggedIn ? 'check_circle' : 'account_circle' }}</mat-icon>
+    <mat-icon>{{ datapoolAuthService.isLoggedIn ? 'check_circle' : 'account_circle' }}</mat-icon>
   </button>
 
   <button
