+++ src/app/services/datapool-auth.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { DatapoolAuthService } from './datapool-auth.service';
+
+describe('DatapoolAuthService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: DatapoolAuthService = TestBed.get(DatapoolAuthService);
+    expect(service).toBeTruthy();
+  });
+});
