+++ src/app/services/index.ts
@@ -9,3 +9,4 @@ export { BulkDownloadService } from './bulk-download.service';
 export { SearchParamsService } from './search-params.service';
 export { RangeService } from './range.service';
 export { PolygonValidationService } from './polygon-validation.service';
+export { DatapoolAuthService } from './datapool-auth.service';
